refact: Separated blockchain logic into it's own file
